Upgrade cucumber-gherkin to v10
Gherkin has changed it's name to cucumber-gherkin ( https://rubygems.org/gems/cucumber-gherkin ), so this bumps the version and uses the new name.

Fixes to the expected metadata from Gherkin. Tags & Examples are now optional, so .blank? is used to test, where .empty? breaks on nil.

Removed link to rubyforge_project in the gemspec, because it's deprecated.

:pear: @jayzz55