Add "missing step" context when raising error for unimplemented steps
I'd like to turn this on on a project so that CI will catch missing
steps, but for local development the existing error is not very
informative compared to the missing step error.

This patch adds the missing context to the backtrace to make it easy to
see that it's a missing step even with the raise config switched on.